:sparkles: : addPromo Refac

+++ discount-management/discountmanagement-adapter/src/main/java/com/fil/sra/bdd/entity/MarketOperationLeastPriceyEntity.java
@@ -11,6 +11,6 @@ import lombok.Setter;
 @Entity
 public class MarketOperationLeastPriceyEntity extends MarketOperationEntity {
 
-    protected int valueOrPercentReduction;
+    protected double valueOrPercentReduction;
 
 }
